fix: It's not backporting it's porting over
Backporting is when you take features from a newer version of software and port them to an older version of the software.

The implication that was made here originally was that Portal 2 is newer than CSGO, which is very incorrect given CSGO released *after* Portal 2, and has it's engine branch branching off of Portal 2's.

As such the correct phrasing would be to say that the features were ported over, not backported.